Moving test files to appropriate crates.
