--- crates/swc/tests/fixture/issue-2101/case1/input/.swcrc
@@ -1,26 +1,0 @@
-{
-  "sourceMaps": true,
-  "jsc": {
-    "parser": {
-      "syntax": "typescript",
-      "tsx": true,
-      "dynamicImport": true
-    },
-    "transform": {
-      "react": {
-        "runtime": "automatic",
-        "refresh": true
-      },
-      "hidden": {
-        "jest": true
-      }
-    }
-  },
-  "module": {
-    "type": "commonjs"
-  },
-  "env": {
-    "coreJs": 3,
-    "mode": "usage"
-  }
-}
