--- crates/swc/tests/fixture/issue-2101/case1/output/index.js
@@ -1,50 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", {
-    value: true
-});
-var _exportNames = {
-  render: true
-};
-Object.defineProperty(exports, "render", {
-    enumerable: true,
-    get: function() {
-        return _customRender.customRender;
-    }
-});
-var _customRender = require("./customRender");
-var _react = _interopRequireWildcard(require("@testing-library/react"));
-Object.keys(_react).forEach(function(key) {
-    if (key === "default" || key === "__esModule") return;
-    if (Object.prototype.hasOwnProperty.call(_exportNames, key)) return;
-    if (key in exports && exports[key] === _react[key]) return;
-    Object.defineProperty(exports, key, {
-        enumerable: true,
-        get: function() {
-            return _react[key];
-        }
-    });
-});
-function _interopRequireWildcard(obj) {
-    if (obj && obj.__esModule) {
-        return obj;
-    } else {
-        var newObj = {
-        };
-        if (obj != null) {
-            for(var key in obj){
-                if (Object.prototype.hasOwnProperty.call(obj, key)) {
-                    var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {
-                    };
-                    if (desc.get || desc.set) {
-                        Object.defineProperty(newObj, key, desc);
-                    } else {
-                        newObj[key] = obj[key];
-                    }
-                }
-            }
-        }
-        newObj.default = obj;
-        return newObj;
-    }
-}
-//# sourceMappingURL=index.js.map
