+++ crates/swc_ecma_transforms_module/tests/fixture/commonjs/issue-2101/input.js
